removed analyze_structure
